+++ CHANGES.rst
@@ -20,7 +20,7 @@ development at the same time, such as 4.5.x and 5.0.
 Unreleased
 ----------
 
-Nothing yet.
+- Debug: added ``pybehave`` to the list of :ref:`cmd_run_debug` options.
 
 
 .. _changes_631: